requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Data handling
pyyaml>=6.0
toml>=0.10.0
requests>=2.27.0

# Repository management
gitpython>=3.1.0

# Text processing
nltk>=3.6.0
wordcloud>=1.8.0

# Network analysis
networkx>=2.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional dependencies for advanced features
# Uncomment if needed
# plotly>=5.5.0
# ipywidgets>=7.6.0
